+++ src/apis/constants.ts
@@ -1,0 +1,9 @@
+export const API_BASE_URL = '/api';
+
+export const API_ENDPOINT = {
+  PRODUCT: '/products',
+  REVIEW: '/reviews',
+  CATEGORY: '/categories',
+  LOGIN: '/login',
+  LOGGED_CHECK: '/logged-check',
+};
