+++ src/constants/index.ts
@@ -1,5 +1,3 @@
-export const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
-
 export const ROUTE = {
   HOME: '/',
   PRODUCT: '/products',
